+++ writer/AbstractMonitoringWriter.py
